chore: Use long in enums and introduce better field finding.

+++ fitfile4j-parser/src/main/java/com/meistermeier/fitfile4j/names/ANT_CHANNEL_ID.java
@@ -4,11 +4,11 @@ package com.meistermeier.fitfile4j.names;
 public enum ANT_CHANNEL_ID {
   _ANT_EXTENDED_DEVICE_NUMBER_UPPER_NIBBLE(4026531840L, "ant_extended_device_number_upper_nibble"),
 
-  _ANT_TRANSMISSION_TYPE_LOWER_NIBBLE(251658240, "ant_transmission_type_lower_nibble"),
+  _ANT_TRANSMISSION_TYPE_LOWER_NIBBLE(251658240L, "ant_transmission_type_lower_nibble"),
 
-  _ANT_DEVICE_TYPE(16711680, "ant_device_type"),
+  _ANT_DEVICE_TYPE(16711680L, "ant_device_type"),
 
-  _ANT_DEVICE_NUMBER(65535, "ant_device_number");
+  _ANT_DEVICE_NUMBER(65535L, "ant_device_number");
 
   private final long messageNumber;
 
@@ -19,7 +19,7 @@ public enum ANT_CHANNEL_ID {
     this.messageName = messageName;
   }
 
-  public static ANT_CHANNEL_ID findById(int messageNumber) {
+  public static ANT_CHANNEL_ID findById(long messageNumber) {
     for (ANT_CHANNEL_ID name: ANT_CHANNEL_ID.values()) {
     	if (name.messageNumber == messageNumber) {
     		return name;
